+++ src/components/pages/index.ts
@@ -1,5 +1,6 @@
 export * from "./home";
 export * from "./artist";
+export * from "./song";
 export * from "./search-artists";
 export * from "./search-songs";
 export * from "./ranking-artists";
